+++ src/rust/bitbox02-rust/src/hww/api/ethereum/sign.rs
@@ -95,10 +95,8 @@ async fn verify_erc20_transaction(
     erc20_recipient: [u8; 20],
     erc20_value: BigUint,
 ) -> Result<(), Error> {
-    let erc20_params = bitbox02::app_eth::erc20_params_get(
-        request.coin as _,
-        parse_recipient(&request.recipient)?,
-    );
+    let erc20_params =
+        bitbox02::app_eth::erc20_params_get(params.chain_id, parse_recipient(&request.recipient)?);
     let formatted_fee = parse_fee(request, params).format();
     let recipient_address = super::address::from_pubkey_hash(&erc20_recipient);
     let (formatted_value, formatted_total) = match erc20_params {
@@ -186,7 +184,7 @@ async fn verify_standard_transaction(
 /// Verify and sign an Ethereum transaction.
 pub async fn process(request: &pb::EthSignRequest) -> Result<Response, Error> {
     let coin = pb::EthCoin::from_i32(request.coin).ok_or(Error::InvalidInput)?;
-    let params = super::params::get(coin).ok_or(Error::InvalidInput)?;
+    let params = super::params::get(coin, request.chain_id).ok_or(Error::InvalidInput)?;
 
     if !super::keypath::is_valid_keypath_address(&request.keypath) {
         return Err(Error::InvalidInput);
@@ -378,6 +376,7 @@ mod tests {
                 value: b"\x07\x5c\xf1\x25\x9e\x9c\x40\x00".to_vec(),
                 data: b"".to_vec(),
                 host_nonce_commitment: None,
+                chain_id: 0,
             })),
             Ok(Response::Sign(pb::EthSignResponse {
                 signature: b"\xc3\xae\x24\xc1\x67\xe2\x16\xcf\xb7\x5c\x72\xb5\xe0\x3e\xf9\x7a\xcc\x2b\x60\x7f\x3a\xcf\x63\x86\x5f\x80\x96\x0f\x76\xf6\x56\x47\x0f\x8e\x23\xf1\xd2\x78\x8f\xb0\x07\x0e\x28\xc2\xa5\xc8\xaa\xf1\x5b\x5d\xbf\x30\xb4\x09\x07\xff\x6c\x50\x68\xfd\xcb\xc1\x1a\x2d\x00"
@@ -432,6 +431,7 @@ mod tests {
             value: b"\x07\x5c\xf1\x25\x9e\x9c\x40\x00".to_vec(),
             data: b"".to_vec(),
             host_nonce_commitment: None,
+            chain_id: 0,
         }))
         .unwrap();
         assert_eq!(unsafe { CONFIRM_COUNTER }, 1);
@@ -482,6 +482,7 @@ mod tests {
                 value: b"\x07\x5c\xf1\x25\x9e\x9c\x40\x00".to_vec(),
                 data: b"foo bar".to_vec(),
                 host_nonce_commitment: None,
+                chain_id: 0,
             })),
             Ok(Response::Sign(pb::EthSignResponse {
                 signature: b"\x7d\x3f\x37\x13\xe3\xcf\x10\x82\x79\x1d\x5c\x0f\xc6\x8e\xc2\x9e\xaf\xf5\xe1\xee\x84\x67\xa8\xec\x54\x7d\xc7\x96\xe8\x5a\x79\x04\x2b\x7c\x01\x69\x2f\xb7\x2f\x55\x76\xab\x50\xdc\xaa\x62\x1a\xd1\xee\xab\xd9\x97\x59\x73\xb8\x62\x56\xf4\x0c\x6f\x85\x50\xef\x44\x00"
@@ -512,7 +513,7 @@ mod tests {
         mock_unlocked();
         assert_eq!(
             block_on(process(&pb::EthSignRequest {
-                coin: pb::EthCoin::Eth as _,
+                coin: pb::EthCoin::RinkebyEth as _, // ignored because chain_id > 0
                 keypath: KEYPATH.to_vec(),
                 nonce: b"\x23\x67".to_vec(),
                 gas_price: b"\x02\x7a\xca\x1a\x80".to_vec(),
@@ -521,6 +522,7 @@ mod tests {
                 value: b"".to_vec(),
                 data: b"\xa9\x05\x9c\xbb\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xe6\xce\x0a\x09\x2a\x99\x70\x0c\xd4\xcc\xcc\xbb\x1f\xed\xc3\x9c\xf5\x3e\x63\x30\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x03\x65\xc0\x40".to_vec(),
                 host_nonce_commitment: None,
+                chain_id: 1,
             })),
             Ok(Response::Sign(pb::EthSignResponse {
                 signature: b"\x67\x4e\x9a\x01\x70\xee\xe0\xca\x8c\x40\x6e\xc9\xa7\xdf\x2e\x3a\x6b\xdd\x17\x9c\xf6\x93\x85\x80\x0e\x1f\xd3\x78\xe7\xcf\xb1\x9c\x4d\x55\x16\x2c\x54\x7b\x04\xd1\x81\x8e\x43\x90\x16\x91\xae\xc9\x88\xef\x75\xcd\x67\xd9\xbb\x30\x1d\x14\x90\x2f\xd6\xe6\x92\x92\x01"
@@ -559,6 +561,7 @@ mod tests {
                 value: b"".to_vec(),
                 data: b"\xa9\x05\x9c\xbb\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x85\x7b\x3d\x96\x9e\xac\xb7\x75\xa9\xf7\x9c\xab\xc6\x2e\xc4\xbb\x1d\x1c\xd6\x0e\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x98\xa6\x3c\xbe\xb8\x59\xd0\x27\xb0".to_vec(),
                 host_nonce_commitment: None,
+                chain_id: 0,
             })),
             Ok(Response::Sign(pb::EthSignResponse {
                 signature: b"\xec\x6e\x53\x0c\x8e\xe2\x54\x34\xfc\x44\x0e\x9a\xc0\xf8\x88\xe9\xc6\x3c\xf0\x7e\xbc\xf1\xc2\xf8\xa8\x3e\x2e\x8c\x39\x83\x2c\x55\x15\x12\x71\x6f\x6e\x1a\x8b\x66\xce\x38\x11\xa7\x26\xbc\xb2\x44\x66\x4e\xf2\x6f\x98\xee\x35\xc0\xc9\xdb\x4c\xaa\xb0\x73\x98\x56\x00"
@@ -581,6 +584,7 @@ mod tests {
             value: b"\x07\x5c\xf1\x25\x9e\x9c\x40\x00".to_vec(),
             data: b"".to_vec(),
             host_nonce_commitment: None,
+            chain_id: 0,
         };
 
         {